+++ trufflesqueak/de.hpi.swa.trufflesqueak/src/de/hpi/swa/trufflesqueak/model/NativeObject.java
@@ -71,7 +71,7 @@ public class NativeObject extends SqueakObject implements TruffleObject {
 
     @Override
     public void atput0(int index, BaseSqueakObject object) throws UnwrappingError {
-        setNativeAt0(index, object.unwrapInt());
+        setNativeAt0(index, (int) object.unwrapInt());
     }
 
     public void atput0(int index, int value) {
